requirements.txt
flask==3.1.0
gunicorn==20.1.0
requests==2.31.0
python-dotenv==1.0.0